+++ truffle/com.oracle.truffle.object.dsl.processor/src/com/oracle/truffle/object/dsl/processor/LayoutParser.java
@@ -62,6 +62,8 @@ public class LayoutParser {
     private boolean hasObjectGuard;
     private boolean hasDynamicObjectGuard;
     private boolean hasShapeProperties;
+    private boolean hasCreate;
+    private boolean hasBuilder;
     private final List<String> constructorProperties = new ArrayList<>();
     private final Map<String, PropertyBuilder> properties = new HashMap<>();
     private List<ImplicitCast> implicitCasts = new ArrayList<>();
@@ -134,6 +136,8 @@ public class LayoutParser {
                     // Handled above
                 } else if (simpleName.equals("create" + name)) {
                     parseConstructor((ExecutableElement) element);
+                } else if (simpleName.equals("build")) {
+                    parseBuilder((ExecutableElement) element);
                 } else if (simpleName.equals("is" + name)) {
                     parseGuard((ExecutableElement) element);
                 } else if (simpleName.startsWith("getAndSet")) {
@@ -202,6 +206,9 @@ public class LayoutParser {
     }
 
     private void parseConstructor(ExecutableElement methodElement) {
+        hasCreate = true;
+        checkCreateAndBuilder(methodElement);
+
         List<? extends VariableElement> parameters = methodElement.getParameters();
 
         if (hasShapeProperties || (superLayout != null && superLayout.hasShapeProperties())) {
@@ -230,6 +237,35 @@ public class LayoutParser {
             checkSharedParameters(methodElement, parameters, superProperties);
         }
 
+        addConstructorProperties(methodElement, parameters);
+    }
+
+    private void parseBuilder(ExecutableElement methodElement) {
+        hasBuilder = true;
+        checkCreateAndBuilder(methodElement);
+
+        List<? extends VariableElement> parameters = methodElement.getParameters();
+
+        if (!isSameType(methodElement.getReturnType(), Object[].class)) {
+            processor.reportError(methodElement, "build() must have Object[] for return type");
+        }
+
+        if (superLayout != null) {
+            final List<PropertyModel> superProperties = superLayout.getAllInstanceProperties();
+            checkSharedParameters(methodElement, parameters, superProperties);
+        }
+
+        addConstructorProperties(methodElement, parameters);
+    }
+
+    private void checkCreateAndBuilder(ExecutableElement methodElement) {
+        if (hasCreate && hasBuilder) {
+            processor.reportError(methodElement, "Only one of create<Layout>() or build() may be specified.");
+            return;
+        }
+    }
+
+    private void addConstructorProperties(ExecutableElement methodElement, List<? extends VariableElement> parameters) {
         for (VariableElement element : parameters) {
             final String parameterName = element.getSimpleName().toString();
 
@@ -237,10 +273,14 @@ public class LayoutParser {
                 processor.reportError(methodElement, "Factory is a confusing name for a property");
             }
 
-            constructorProperties.add(parameterName);
-            final PropertyBuilder property = getProperty(parameterName);
-            setPropertyType(element, property, element.asType());
-            parseConstructorParameterAnnotations(property, element);
+            if (constructorProperties.contains(parameterName)) {
+                processor.reportError(methodElement, "The property %s is duplicated");
+            } else {
+                constructorProperties.add(parameterName);
+                final PropertyBuilder property = getProperty(parameterName);
+                setPropertyType(element, property, element.asType());
+                parseConstructorParameterAnnotations(property, element);
+            }
         }
     }
 
@@ -550,7 +590,7 @@ public class LayoutParser {
 
     public LayoutModel build() {
         return new LayoutModel(objectTypeSuperclass, superLayout, name, packageName, hasObjectTypeGuard, hasObjectGuard,
-                        hasDynamicObjectGuard, buildProperties(), interfaceFullName, implicitCasts);
+                        hasDynamicObjectGuard, hasBuilder, buildProperties(), interfaceFullName, implicitCasts);
     }
 
     private List<PropertyModel> buildProperties() {